Renomeado stories de Blocks e Templates
